requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
PyMySQL==1.0.2
python-dotenv==0.17.1
SQLAlchemy==1.4.15
Werkzeug==2.0.0
WTForms==2.3.3
